+++ src/main/java/codesquad/Main.java
@@ -6,9 +6,14 @@ import codesquad.handler.ResourceHandlerAdapter;
 import codesquad.http.HttpMethod;
 import codesquad.http.HttpResponseSerializer;
 import codesquad.model.User;
+import codesquad.model.business.LoginUserLogic;
 import codesquad.model.business.RegisterUserLogic;
 import codesquad.processor.*;
+import codesquad.processor.argumentresolver.ArgumentResolver;
+import codesquad.processor.argumentresolver.LoginArgumentResolver;
+import codesquad.processor.argumentresolver.RegisterArgumentResolver;
 import codesquad.server.ServerInitializer;
+import codesquad.web.user.LoginRequest;
 import codesquad.web.user.RegisterRequest;
 
 import java.util.ArrayList;
