+++ src/main/java/codesquad/processor/argumentresolver/RegisterArgumentResolver.java
@@ -1,11 +1,10 @@
-package codesquad.processor;
+package codesquad.processor.argumentresolver;
 
+import codesquad.helper.RequestBodyParseHelper;
 import codesquad.http.HttpRequest;
 import codesquad.web.user.RegisterRequest;
 
-import java.util.Arrays;
 import java.util.Map;
-import java.util.stream.Collectors;
 
 public class RegisterArgumentResolver implements ArgumentResolver<RegisterRequest> {
 
@@ -14,7 +13,7 @@ public class RegisterArgumentResolver implements ArgumentResolver<RegisterReques
 
         //아직 PathVariable을 처리하는 방법은 생각하지 않았음
         // 임시로 RegisterRequest를 만들기, TODO 필요에 따라 다양하게 만들 수 있게 수정
-        Map<String, String> bodyParameters = bodyParameters(httpRequest.getBody());
+        Map<String, String> bodyParameters = RequestBodyParseHelper.bodyParameters(httpRequest.getBody());
 
         final String email = bodyParameters.get("email");
         final String userId = bodyParameters.get("userId");
@@ -26,20 +25,6 @@ public class RegisterArgumentResolver implements ArgumentResolver<RegisterReques
         }
 
         return new RegisterRequest(email, userId, password, name);
-    }
-
-    private Map<String, String> bodyParameters(String bodyString) {
-        return Arrays.stream(bodyString.split("&"))
-                .map(s -> {
-                    String[] split = s.split("=");
-                    String key = split[0];
-                    String value = split.length > 1 ? split[1] : "";
-
-                    return Map.entry(key, value);
-                })
-
-                .collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue));
-
     }
 
 }
